pyproject.toml
[project]
name = "notesound"
version = "0.1.0"
description = "MVP API for recognizing melodies from sheet music photos."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "alembic>=1.14.0",
    "defusedxml>=0.7.1",
    "fastapi>=0.115.0",
    "homr>=0.6.2",
    "music21>=9.3.0",
    "numpy>=2.1.0",
    "opencv-python-headless>=4.10.0.84",
    "pillow>=11.0.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "python-json-logger>=2.0.7",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "app",
    "app.api",
    "app.core",
    "app.domain",
    "app.repositories",
    "app.services.preprocess",
    "app.services.omr",
    "app.services.music",
    "app.workers",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

